SL-51: refactor store, quest page challenge 1 working
