+++ src/helpers/Sort/sortByState.js
@@ -1,7 +1,7 @@
 const sortByState = (arr) => {
   let filterArr = [...arr];
   return filterArr.sort((a, b) =>
-    a.checked && !b.checked ? -1 : !a.checked && b.checked ? 1 : 0,
+    a.completed && !b.completed ? -1 : !a.completed && b.completed ? 1 : 0,
   );
 };
