--- src/helpers/updateQuests.js
@@ -1,5 +1,0 @@
-const openCongratsModal = (modalId) => {
-  document.getElementById(modalId).showModal();
-};
-
-export { openCongratsModal };
