+++ src/views/LandingPage.vue
@@ -1,5 +1,5 @@
 <template>
-  <ParentHeader @click="showCongratsModal"> SayLess </ParentHeader>
+  <ParentHeader> SayLess </ParentHeader>
   <div class="flex flex-col md:flex-row justify-center md:gap-4">
     <div
       id="divOne"
@@ -90,9 +90,6 @@ export default {
     pageLoadAnimation(divIds);
   },
   methods: {
-    showCongratsModal() {
-      openCongratsModal('congrats_1');
-    },
     navigateInventory() {
       this.$router.push('/inventory');
     },
