+++ src/views/QuestPage.vue
@@ -2,16 +2,18 @@
   <div class="flex justify-center items-center mt-16 md:mt-40">
     <div class="relative card w-60 md:w-48 glass shadow-md">
       <div class="card-body gap-0 p-4 pt-6 items-center text-center">
-        <h1 class="card-title text-pink">Quest 1</h1>
-        <p class="text-sm text-white-light">complete challenges</p>
+        <h1 class="card-title text-pink">{{ getQuestData.questName }}</h1>
+        <p class="text-sm text-white-light">Complete Challenges</p>
         <div class="absolute -top-3.5">
-          <QuestStars />
+          <QuestStars :challenge-status="getChallengesChecked" />
         </div>
       </div>
     </div>
   </div>
-
-  <GraphicComponent class="m-auto mt-12" />
+  <GraphicComponent
+    class="m-auto mt-12"
+    :challenge-status="getChallengeStatus"
+  />
   <ChallengesContainer />
 </template>
 
@@ -19,6 +21,7 @@
 import ChallengesContainer from '@/components/Quest/ChallengesContainer.vue';
 import QuestStars from '@/components/Quest/QuestStars.vue';
 import GraphicComponent from '@/components/GraphicComponent/GraphicComponent.vue';
+import { mapGetters } from 'vuex';
 
 export default {
   name: 'QuestPage',
@@ -27,10 +30,14 @@ export default {
     QuestStars,
     GraphicComponent,
   },
-  data() {
-    return {
-      email: this.$store.getters.getEmail,
-    };
+  computed: {
+    ...mapGetters(['getQuestData']),
+    getChallengesChecked() {
+      return this.getQuestData.challenges;
+    },
+    getChallengeStatus() {
+      return this.getQuestData.challenges[0].status;
+    },
   },
   methods: {
     navigateToForm() {
